bsppp: move BSP class code out to BSP.h/cpp
