+++ src/bsppp/BSP.cpp
@@ -1,4 +1,4 @@
-#include <bsppp/bsppp.h>
+#include <bsppp/BSP.h>
 
 #include <algorithm>
 #include <cstdint>
